address PR feedback
* DRY externalize_header_map
* move runner into submodule of replay
* add more context to the readme (it's not docs, it's a file included with the recordings)
