+++ apollo-router/src/plugins/coprocessor/mod.rs
@@ -37,6 +37,7 @@ use crate::plugin::Plugin;
 use crate::plugin::PluginInit;
 use crate::register_plugin;
 use crate::services;
+use crate::services::external::externalize_header_map;
 use crate::services::external::Control;
 use crate::services::external::Externalizable;
 use crate::services::external::PipelineStep;
@@ -1140,19 +1141,6 @@ fn validate_coprocessor_output<T>(
         )));
     }
     Ok(())
-}
-
-/// Convert a HeaderMap into a HashMap
-pub(crate) fn externalize_header_map(
-    input: &HeaderMap<HeaderValue>,
-) -> Result<HashMap<String, Vec<String>>, BoxError> {
-    let mut output = HashMap::new();
-    for (k, v) in input {
-        let k = k.as_str().to_owned();
-        let v = String::from_utf8(v.as_bytes().to_vec()).map_err(|e| e.to_string())?;
-        output.entry(k).or_insert_with(Vec::new).push(v)
-    }
-    Ok(output)
 }
 
 /// Convert a HashMap into a HeaderMap
