+++ apollo-router/src/plugins/record_replay/mod.rs
@@ -1,4 +1,3 @@
 mod record;
 mod recording;
 mod replay;
-mod runner;
