+++ apollo-router/src/plugins/record_replay/recording-readme.md
@@ -1,37 +1,20 @@
 # Apollo Router Request Recordings
 
-To record a request:
-
-- enable the recording plugin in the router:
-
-       ```yaml
-       plugins:
-        experimental.record:
-          enabled: true
-          storage_path: /tmp/recordings/
-       ```
-
-- add `x-apollo-router-record: 1` as a request header:
-
-      ```sh
-      curl http://localhost:4000/ \
-        -H 'content-type: application/json' \
-        -H 'x-apollo-router-record: 1' \
-        -d '{"query": "{ me { id } }"}'
-      ```
+The files in this directory contain recordings of GraphQL requests and responses to Apollo Router. They may contain private or sensitive data, so please review before sharing.
 
 ## Sharing with Apollo
 
-Please include a copy of your router configuration.
+Please include a copy of your router configuration for accurate reproductions.
+
+**Do not share recordings publicly without reviewing the recorded data!**
 
 ## Security considerations
 
 Recordings may capture:
 
 - Authentication tokens in HTTP headers
+- Private data in request operations or variables
 - Private data in responses
-
-**Do not share recordings publicly without reviewing the recorded data!**
 
 ## Replay a recording
 
@@ -40,5 +23,5 @@ Inside the [Apollo Router codebase](https://www.github.com/apollographql/router)
 ```sh
 RECORDING_FILE=/tmp/recordings/Query-1698253358.json \
   cargo test --package apollo-router --lib \
-  -- plugins::record_replay::runner::tests::replay_recording --exact --nocapture
+  -- plugins::record_replay::replay::tests::replay_recording --exact --nocapture
 ```
